requirements.txt
aiogram==3.17.0
yt-dlp
python-dotenv